+++ packages/bugsnag_flutter_performance/lib/src/client.dart
@@ -10,6 +10,7 @@ import 'package:bugsnag_flutter_performance/src/instrumentation/navigation/navig
 import 'package:bugsnag_flutter_performance/src/instrumentation/view_load/measured_widget_callbacks.dart';
 import 'package:bugsnag_flutter_performance/src/instrumentation/view_load/view_load_instrumentation.dart';
 import 'package:bugsnag_flutter_performance/src/span_attributes.dart';
+import 'package:bugsnag_flutter_performance/src/span_attributes_limits.dart';
 import 'package:bugsnag_flutter_performance/src/span_context.dart';
 import 'package:bugsnag_flutter_performance/src/uploader/package_builder.dart';
 import 'package:bugsnag_flutter_performance/src/uploader/retry_queue.dart';
@@ -141,6 +142,9 @@ class BugsnagPerformanceClientImpl implements BugsnagPerformanceClient {
     String? serviceName,
     String? appVersion,
     double? samplingProbability,
+    int? attributeCountLimit,
+    int? attributeStringValueLimit,
+    int? attributeArrayLengthLimit,
     List<OnSpanEndCallback>? onSpanEndCallbacks,
   }) async {
     if (!_isEnabledOnCurrentPlatform()) {
@@ -151,6 +155,11 @@ class BugsnagPerformanceClientImpl implements BugsnagPerformanceClient {
     WidgetsFlutterBinding.ensureInitialized();
     _networkRequestCallback = networkRequestCallback;
     _onSpanEndCallbacks = onSpanEndCallbacks ?? [];
+    BugsnagPerformanceSpanImpl.globalAttributeCountLimit =
+        SpanAttributesLimits.limitValue(
+      type: SpanAttributesLimitType.attributeCountLimit,
+      providedValue: attributeCountLimit,
+    );
     configuration = BugsnagPerformanceConfiguration(
       apiKey: apiKey,
       endpoint: endpoint ?? Uri.parse(_defaultEndpoint(apiKey)),
@@ -160,7 +169,21 @@ class BugsnagPerformanceClientImpl implements BugsnagPerformanceClient {
       serviceName: serviceName,
       appVersion: appVersion,
       samplingProbability: samplingProbability,
+      attributeCountLimit: SpanAttributesLimits.limitValue(
+        type: SpanAttributesLimitType.attributeCountLimit,
+        providedValue: attributeCountLimit,
+      ),
+      attributeStringValueLimit: SpanAttributesLimits.limitValue(
+        type: SpanAttributesLimitType.stringValueLimit,
+        providedValue: attributeStringValueLimit,
+      ),
+      attributeArrayLengthLimit: SpanAttributesLimits.limitValue(
+        type: SpanAttributesLimitType.arrayLengthLimit,
+        providedValue: attributeArrayLengthLimit,
+      ),
     );
+    BugsnagPerformanceSpanImpl.globalAttributeCountLimit =
+        configuration!.attributeCountLimit;
     _packageBuilder.setConfig(configuration);
     _initialExtraConfig.forEach((key, value) {
       setExtraConfig(key, value);
@@ -219,6 +242,7 @@ class BugsnagPerformanceClientImpl implements BugsnagPerformanceClient {
       parentSpanId: parent?.spanId,
       traceId: parent?.traceId,
       attributes: attributes,
+      attributeCountLimit: configuration?.attributeCountLimit,
     );
     span.clock = _clock;
     if (configuration != null) {