+++ packages/bugsnag_flutter_performance/lib/src/span_attributes.dart
@@ -1,3 +1,17 @@
+import 'package:bugsnag_flutter_performance/src/configuration.dart';
+import 'package:bugsnag_flutter_performance/src/span_attributes_limits.dart';
+import 'package:flutter/foundation.dart';
+
+class BugsnagPerformanceSpanAttributesEncodingResult {
+  final dynamic jsonValue;
+  final int droppedAttributesCount;
+
+  BugsnagPerformanceSpanAttributesEncodingResult({
+    required this.jsonValue,
+    required this.droppedAttributesCount,
+  });
+}
+
 class BugsnagPerformanceSpanAttributes {
   final Map<String, dynamic> attributes = {};
 
@@ -51,16 +65,29 @@ class BugsnagPerformanceSpanAttributes {
     return attributes['bugsnag.sampling.p'];
   }
 
-  dynamic toJson() {
-    return attributes.entries.map((entry) {
-      final key = entry.key;
-      final value = entry.value;
+  int get count => attributes.length;
 
-      return {
-        'key': key,
-        'value': _encodeAttributeValue(value),
-      };
-    }).toList();
+  bool hasAttribute(String key) {
+    return attributes.containsKey(key);
+  }
+
+  BugsnagPerformanceSpanAttributesEncodingResult toJson({
+    BugsnagPerformanceConfiguration? config,
+  }) {
+    final attributeKeyLengthLimit = SpanAttributesLimits.limitValue(
+        type: SpanAttributesLimitType.keyLengthLimit);
+    final attributeStringValueLimit = config?.attributeStringValueLimit ??
+        SpanAttributesLimits.limitValue(
+            type: SpanAttributesLimitType.stringValueLimit);
+    final attributeArrayLengthLimit = config?.attributeArrayLengthLimit ??
+        SpanAttributesLimits.limitValue(
+            type: SpanAttributesLimitType.arrayLengthLimit);
+
+    return BugsnagPerformanceSpanAttributesEncoder(
+      attributeKeyLengthLimit: attributeKeyLengthLimit,
+      attributeStringValueLimit: attributeStringValueLimit,
+      attributeArrayLengthLimit: attributeArrayLengthLimit,
+    ).toJson(attributes);
   }
 
   factory BugsnagPerformanceSpanAttributes.fromJson(
@@ -76,21 +103,6 @@ class BugsnagPerformanceSpanAttributes {
     }
     return BugsnagPerformanceSpanAttributes(additionalAttributes: attributes);
   }
-
-  static const _typeMap = {
-    bool: 'boolValue',
-    double: 'doubleValue',
-    int: 'intValue',
-    String: 'stringValue'
-  };
-
-  static const _listTypeMap = {
-    List<bool>: 'arrayValue',
-    List<double>: 'arrayValue',
-    List<int>: 'arrayValue',
-    List<String>: 'arrayValue',
-    List<Object>: 'arrayValue'
-  };
 
   static dynamic _decodeAttributeValue(Map<String, dynamic> valueMap) {
     if (valueMap.containsKey('stringValue')) {
@@ -116,32 +128,111 @@ class BugsnagPerformanceSpanAttributes {
     }
     return null;
   }
+}
 
-  dynamic _encodeAttributeValue(dynamic value) {
+class BugsnagPerformanceSpanAttributesEncoder {
+  int attributeKeyLengthLimit;
+  int attributeStringValueLimit;
+  int attributeArrayLengthLimit;
+
+  BugsnagPerformanceSpanAttributesEncoder({
+    required this.attributeKeyLengthLimit,
+    required this.attributeStringValueLimit,
+    required this.attributeArrayLengthLimit,
+  });
+
+  static const _typeMap = {
+    bool: 'boolValue',
+    double: 'doubleValue',
+    int: 'intValue',
+    String: 'stringValue'
+  };
+
+  static const _listTypeMap = {
+    List<bool>: 'arrayValue',
+    List<double>: 'arrayValue',
+    List<int>: 'arrayValue',
+    List<String>: 'arrayValue',
+    List<Object>: 'arrayValue'
+  };
+
+  BugsnagPerformanceSpanAttributesEncodingResult toJson(
+    Map<String, dynamic> attributes,
+  ) {
+    final validKeys =
+        attributes.keys.where((key) => key.length <= attributeKeyLengthLimit);
+
+    final jsonValue = validKeys.map((key) {
+      final value = attributes[key];
+
+      return {
+        'key': key,
+        'value': _encodeAttributeValue(key, value),
+      };
+    }).toList();
+
+    return BugsnagPerformanceSpanAttributesEncodingResult(
+      jsonValue: jsonValue,
+      droppedAttributesCount: attributes.length - validKeys.length,
+    );
+  }
+
+  dynamic _encodeAttributeValue(
+    String key,
+    dynamic value,
+  ) {
     final runtimeType = value.runtimeType;
     String valueType =
         _typeMap[runtimeType] ?? _listTypeMap[runtimeType] ?? 'stringValue';
-    dynamic formattedValue = _formattedValue(value);
+    dynamic formattedValue = _formattedValue(key, value);
     return {valueType: formattedValue};
   }
 
-  dynamic _formattedValue(dynamic value) {
+  dynamic _formattedValue(
+    String key,
+    dynamic value,
+  ) {
+    if (value is String) {
+      if (value.length > attributeStringValueLimit) {
+        if (kDebugMode) {
+          print(
+              'The value for span attribute "$key" was truncated as it exceeds the $attributeStringValueLimit attribute limit set by AttributeStringValueLimit');
+        }
+        return '${value.substring(0, attributeStringValueLimit)}*** ${value.length - attributeStringValueLimit} CHARS TRUNCATED';
+      }
+      return value;
+    }
     if (value is int) {
       return value.toString();
     }
     if (value is List) {
-      return _encodeListAttributeValue(value);
+      return _encodeListAttributeValue(key, value);
     }
     return value;
   }
 
-  dynamic _encodeListAttributeValue(List listAttribute) {
+  dynamic _encodeListAttributeValue(
+    String key,
+    List listAttribute,
+  ) {
     List result = [];
-    for (var value in listAttribute) {
+    for (var (index, value) in listAttribute.indexed) {
       final runtimeType = value.runtimeType;
       if (_typeMap[runtimeType] != null) {
-        result.add(_encodeAttributeValue(value));
+        result.add(_encodeAttributeValue(key, value));
+      } else {
+        if (kDebugMode) {
+          print(
+              'The element at index $index for span attribute "$key" was excluded as its type is not valid.');
+        }
       }
+    }
+    if (result.length > attributeArrayLengthLimit) {
+      if (kDebugMode) {
+        print(
+            'The value for span attribute "$key" was truncated as it exceeds the $attributeArrayLengthLimit attribute limit set by AttributeArrayLengthLimit');
+      }
+      result = result.sublist(0, attributeArrayLengthLimit).toList();
     }
     return {'values': result};
   }